Split Pragha between the application and a library to link the plugins.
This results in a true plugin system, where other developers can implement 3rd party plugins for extend Pragha.

I wrote a small sample to test how it works:
 * https://github.com/pragha-music-player/pragha-sample-plugin

...that implements the option to start a CD burner application with the current playlist.

Dont remove old widgets when remove extra button on toolbar.. and append an api to it.
