+++ plugins/ampache/Makefile.am
@@ -13,30 +13,14 @@ libpampache_la_SOURCES =		\
 
 libpampache_la_LDFLAGS = $(PLUGIN_LIBTOOL_FLAGS)
 
-libpampache_la_CFLAGS =						\
-	$(GRILO_NET_CFLAGS)						\
-	$(GSTREAMER_CFLAGS)						\
-	$(GSTREAMER_BASE_CFLAGS)				\
-	$(GLIB_CFLAGS) 							\
-	$(GIO_CFLAGS) 							\
-	$(GTK_CFLAGS) 							\
-	$(SQLITE3_CFLAGS) 						\
-	$(TAGLIB_C_CFLAGS) 						\
-	$(LIBPEAS_CFLAGS) 						\
-	$(LIBPEAS_GTK_CFLAGS) 					\
-	-I$(top_srcdir)/src/include 			\
-	-Wall
+libpampache_la_CFLAGS = \
+	$(PRAGHA_CFLAGS) \
+	$(GRILO_NET_CFLAGS)
 
-libpampache_la_LIBADD =						\
-	$(GRILO_NET_LIBS) 						\
-	$(GSTREAMER_LIBS)						\
-	$(GLIB_LIBS) 							\
-	$(GIO_LIBS) 							\
-	$(GTK_LIBS) 							\
-	$(SQLITE3_LIBS)							\
-	$(TAGLIB_C_LIBS)						\
-	$(LIBPEAS_LIBS)							\
-	$(LIBPEAS_GTK_LIBS)
+libpampache_la_LIBADD = \
+	$(PRAGHA_LIBS) \
+	$(GRILO_NET_LIBS) \
+	$(top_builddir)/src/libpragha.la
 
 if HAVE_GRILO3
 libpampache_la_CFLAGS += $(GRILO_NET3_CFLAGS)
@@ -45,16 +29,6 @@ endif
 if HAVE_GRILO2
 libpampache_la_CFLAGS += $(GRILO_NET2_CFLAGS)
 libpampache_la_LIBADD += $(GRILO_NET2_LIBS)
-endif
-
-if HAVE_GSTREAMER_AUDIO
-libpampache_la_CFLAGS += $(GSTREAMER_AUDIO_CFLAGS)
-libpampache_la_LIBADD += $(GSTREAMER_AUDIO_LIBS)
-endif
-
-if HAVE_LIBXFCE4UI
-libpampache_la_CFLAGS += $(LIBXFCE4UI_CFLAGS)
-libpampache_la_LIBADD += $(LIBXFCE4UI_LIBS)
 endif
 
 plugin_DATA = ampache.plugin
