+++ plugins/dlna-renderer/Makefile.am
@@ -13,28 +13,12 @@ libpdlnarenderer_la_SOURCES =		\
 
 libpdlnarenderer_la_LDFLAGS = $(PLUGIN_LIBTOOL_FLAGS)
 
-libpdlnarenderer_la_CFLAGS =				\
-	$(GSTREAMER_CFLAGS)						\
-	$(GSTREAMER_BASE_CFLAGS)				\
-	$(GLIB_CFLAGS) 							\
-	$(GIO_CFLAGS) 							\
-	$(GTK_CFLAGS) 							\
-	$(SQLITE3_CFLAGS) 						\
-	$(TAGLIB_C_CFLAGS) 						\
-	$(LIBPEAS_CFLAGS) 						\
-	$(LIBPEAS_GTK_CFLAGS) 					\
-	-I$(top_srcdir)/src/include 			\
-	-Wall
+libpdlnarenderer_la_CFLAGS = \
+	$(PRAGHA_CFLAGS)
 
-libpdlnarenderer_la_LIBADD =				\
-	$(GSTREAMER_LIBS)						\
-	$(GLIB_LIBS) 							\
-	$(GIO_LIBS) 							\
-	$(GTK_LIBS) 							\
-	$(SQLITE3_LIBS)							\
-	$(TAGLIB_C_LIBS)						\
-	$(LIBPEAS_LIBS)							\
-	$(LIBPEAS_GTK_LIBS)
+libpdlnarenderer_la_LIBADD = \
+	$(PRAGHA_LIBS) \
+	$(top_builddir)/src/libpragha.la
 
 if HAVE_GRILO3
 libpdlnarenderer_la_CFLAGS += $(GRILO3_CFLAGS)
@@ -43,16 +27,6 @@ endif
 if HAVE_GRILO2
 libpdlnarenderer_la_CFLAGS += $(GRILO2_CFLAGS)
 libpdlnarenderer_la_LIBADD += $(GRILO2_LIBS)
-endif
-
-if HAVE_GSTREAMER_AUDIO
-libpdlnarenderer_la_CFLAGS += $(GSTREAMER_AUDIO_CFLAGS)
-libpdlnarenderer_la_LIBADD += $(GSTREAMER_AUDIO_LIBS)
-endif
-
-if HAVE_LIBXFCE4UI
-libpdlnarenderer_la_CFLAGS += $(LIBXFCE4UI_CFLAGS)
-libpdlnarenderer_la_LIBADD += $(LIBXFCE4UI_LIBS)
 endif
 
 plugin_DATA = dlna-renderer.plugin
