+++ plugins/lastfm/Makefile.am
@@ -12,43 +12,17 @@ libplastfm_la_SOURCES =      \
 
 libplastfm_la_LDFLAGS = $(PLUGIN_LIBTOOL_FLAGS)
 
-libplastfm_la_CFLAGS =						\
-	$(LIBCLASTFM_CFLAGS)					\
-	$(GSTREAMER_CFLAGS)						\
-	$(GSTREAMER_BASE_CFLAGS)				\
-	$(GLIB_CFLAGS) 							\
-	$(GIO_CFLAGS) 							\
-	$(GTK_CFLAGS) 							\
-	$(SQLITE3_CFLAGS) 						\
-	$(TAGLIB_C_CFLAGS) 						\
-	$(LIBPEAS_CFLAGS) 						\
-	$(LIBPEAS_GTK_CFLAGS) 					\
-	-I$(top_srcdir)/src/include 			\
-	-Wall
+libplastfm_la_CFLAGS = \
+	$(PRAGHA_CFLAGS) \
+	$(LIBCLASTFM_CFLAGS)
 
-libplastfm_la_LIBADD =						\
-	$(LIBCLASTFM_LIBS)						\
-	$(GSTREAMER_LIBS)						\
-	$(GLIB_LIBS) 							\
-	$(GIO_LIBS) 							\
-	$(GTK_LIBS) 							\
-	$(SQLITE3_LIBS)							\
-	$(TAGLIB_C_LIBS)						\
-	$(LIBPEAS_LIBS)							\
-	$(LIBPEAS_GTK_LIBS)
-
-if HAVE_GSTREAMER_AUDIO
-libplastfm_la_CFLAGS += $(GSTREAMER_AUDIO_CFLAGS)
-libplastfm_la_LIBADD += $(GSTREAMER_AUDIO_LIBS)
-endif
-
-if HAVE_LIBXFCE4UI
-libplastfm_la_CFLAGS += $(LIBXFCE4UI_CFLAGS)
-libplastfm_la_LIBADD += $(LIBXFCE4UI_LIBS)
-endif
+libplastfm_la_LIBADD = \
+	$(PRAGHA_LIBS) \
+	$(LIBCLASTFM_LIBS) \
+	$(top_builddir)/src/libpragha.la
 
 pragha-lastfm-menu-ui.h: pragha-lastfm-menu.ui Makefile
-	$(AM_V_GEN) exo-csource --strip-comments --strip-content --static --name=lastfm_menu_ui $< > $@
+	$(AM_V_GEN) xdt-csource --strip-comments --strip-content --static --name=lastfm_menu_ui $< > $@
 
 plugin_DATA = lastfm.plugin
