+++ plugins/mtp/Makefile.am
@@ -18,42 +18,14 @@ libpmtp_la_SOURCES = \
 
 libpmtp_la_LDFLAGS = $(PLUGIN_LIBTOOL_FLAGS)
 
-libpmtp_la_CFLAGS =							\
-	$(GSTREAMER_CFLAGS)						\
-	$(GSTREAMER_BASE_CFLAGS)					\
-	$(GLIB_CFLAGS) 							\
-	$(GIO_CFLAGS) 							\
-	$(GTK_CFLAGS) 							\
-	$(SQLITE3_CFLAGS) 						\
-	$(TAGLIB_C_CFLAGS) 						\
-	$(LIBPEAS_CFLAGS) 						\
-	$(LIBPEAS_GTK_CFLAGS) 						\
-	$(GUDEV_CFLAGS) 						\
-	$(LIBMTP_CFLAGS)						\
-	-I$(top_srcdir)/src/include 					\
-	-Wall
+libpmtp_la_CFLAGS = \
+	$(PRAGHA_CFLAGS) \
+	$(LIBMTP_CFLAGS)
 
-libpmtp_la_LIBADD =							\
-	$(GSTREAMER_LIBS)						\
-	$(GLIB_LIBS) 							\
-	$(GIO_LIBS) 							\
-	$(GTK_LIBS) 							\
-	$(SQLITE3_LIBS)							\
-	$(TAGLIB_C_LIBS)						\
-	$(LIBPEAS_LIBS)							\
-	$(LIBPEAS_GTK_LIBS)						\
-	$(GUDEV_LIBS) 							\
-	$(LIBMTP_LIBS)
-
-if HAVE_GSTREAMER_AUDIO
-libpmtp_la_CFLAGS += $(GSTREAMER_AUDIO_CFLAGS)
-libpmtp_la_LIBADD += $(GSTREAMER_AUDIO_LIBS)
-endif
-
-if HAVE_LIBXFCE4UI
-libpmtp_la_CFLAGS += $(LIBXFCE4UI_CFLAGS)
-libpmtp_la_LIBADD += $(LIBXFCE4UI_LIBS)
-endif
+libpmtp_la_LIBADD = \
+	$(PRAGHA_LIBS) \
+	$(LIBMTP_LIBS) \
+	$(top_builddir)/src/libpragha.la
 
 plugin_DATA = mtp.plugin
