+++ src/libpragha.pc.in
@@ -1,0 +1,11 @@
+prefix=@prefix@
+libdir=@libdir@/pragha
+includedir=@includedir@/libpragha
+
+Name: libpragha
+Description: Library to write Pragha plugins
+Requires: gstreamer-1.0 gstreamer-base-1.0 gio-2.0 gtk+-3.0 gtk+-3.0 taglib_c libpeas-1.0 libpeas-gtk-1.0
+Version: @VERSION@
+Libs: -L${libdir} -lpragha
+Cflags: -I${includedir}
+
